mcp/memory/requirements.txt
mem0ai>=0.1.0
qdrant-client>=1.9.0
mcp>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
sse-starlette>=2.0.0
